Move xdr parser to a separate file
